docker/clip_encoder/requirements.txt
torch==2.0.1
transformers
flask
gunicorn
pillow
numpy==1.*